+++ MacPacker/Views/WelcomeView.swift
@@ -59,6 +59,7 @@ struct WhatsNewPill: View {
 struct WelcomeWhatsNewView: View {
     var body: some View {
         VStack(alignment: .leading, spacing: 4) {
+            WhatsNewPill(title: "Quick Look extension to preview archives in Finder", type: .feature)
             WhatsNewPill(title: "App crashes when all windows are closed and then the app gets terminated", type: .bug)
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)